Struct pencil::Module

pub struct Module {
    pub name: String,
    pub root_path: String,
    pub static_folder: Option<String>,
    pub static_url_path: Option<String>,
    pub template_folder: Option<String>,
    // some fields omitted
}

Represents a module.

Fields
name: String

The name of the module.

root_path: String

The path where your module locates.

static_folder: Option<String>

The folder with static files that should be served at static_url_path.

static_url_path: Option<String>

The url path for the static files on the web.

template_folder: Option<String>

The folder that contains the templates that should be used for the module.

Methods

impl Module

fn new(name: &str, root_path: &str) -> Module

fn route<M: Into<Matcher>, N: AsRef<[Method]>>(&mut self, rule: M, methods: N, endpoint: &str, view_func: ViewFunc)

The endpoint is automatically prefixed with the module's name.

fn before_request(&mut self, f: BeforeRequestFunc)

Before request for a module. This is only executed before each request that is handled by a view function of that module.

fn before_app_request(&mut self, f: BeforeRequestFunc)

Before request for the app that this module is registered on. This is executed before each request, even if outside of a module.

fn after_request(&mut self, f: AfterRequestFunc)

After request for a module. This is only executed after each request that is handled by a view function of that module.

fn after_app_request(&mut self, f: AfterRequestFunc)

After request for the app that this module is registered on. This is executed after each request, even if outside of a module.

fn teardown_request(&mut self, f: TeardownRequestFunc)

Teardown request for a module. This is only executed when tearing down each request that is handled by a view function of that module.

fn teardown_app_request(&mut self, f: TeardownRequestFunc)

Teardown request for the app that this module is registered on. This is executed when tearing down each request, even if outside of a module.

fn httperrorhandler(&mut self, status_code: u16, f: HTTPErrorHandler)

Registers a http error handler that becomes active for this module only.

fn usererrorhandler(&mut self, error_desc: &str, f: UserErrorHandler)

Registers an user error handler that becomes active for this module only.

fn app_httperrorhandler(&mut self, status_code: u16, f: HTTPErrorHandler)

Registers a http error handler for all requests of the application.

fn app_usererrorhandler(&mut self, error_desc: &str, f: UserErrorHandler)

Registers an user error handler for all requests of the application.

fn register(self, app: &mut Pencil)

Register this module.