Struct pencil::wrappers::Response

pub struct Response {
    pub status_code: u16,
    pub headers: Headers,
    pub body: Option<Box<BodyWrite>>,
}

Response type. It is just one container with a couple of parameters (headers, body, status code etc).

Fields

status_code

The HTTP Status code number

headers

body

Methods

impl Response

fn new<T: 'static + BodyWrite>(body: T) -> Response

Create a Response. By default, the status code is 200 and content type is "text/html; charset=UTF-8". Remember to set content length if necessary. Mostly you should just get a response that is converted from other types, which set the content length automatically. For example:

// Content length is set automatically
let response = Response::from("Hello");

fn new_empty() -> Response

Create an empty response without body.

fn status_name(&self) -> &str

Get status name.

fn content_type(&self) -> Option<&ContentType>

Returns the response content type if available.

fn set_content_type(&mut self, mimetype: &str)

Set response content type. If the mimetype passed is a mimetype starting with text/ or something that needs a charset, the charset(UTF-8) parameter is appended to it.

fn content_length(&self) -> Option<usize>

Returns the response content length if available.

fn set_content_length(&mut self, value: usize)

Set content length.

fn set_cookie(&mut self, cookie: SetCookie)

Sets cookie.